Move `evolve_with_{one,two}` to module `evolution`
